Fix import path for Setup component and refactor joke fetching logic in Headers component
